pyproject.toml
[project]
name = "text-annotation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.12",
    "pydantic>=2.11.5",
    "pyyaml>=6.0.2",
    "sqlalchemy>=2.0.41",
    "uvicorn>=0.34.3",
]
